+++ pysys-examples/internal/testcases/PySys_internal_080/Input/nestedtests/customrunner.py
@@ -1,0 +1,16 @@
+from pysys.baserunner import BaseRunner
+from pysys.constants import *
+from pysys.process.monitor import *
+
+class MyCustomRunner(BaseRunner):
+	
+	def setup(self, **kwargs):
+		super(MyCustomRunner, self).setup(**kwargs)
+		# matches what we put into changelog to document how to go back to old behaviour
+		ProcessMonitorTextFileHandler.setDefaults(
+        [
+           ProcessMonitorKey.DATE_TIME_LEGACY, 
+           ProcessMonitorKey.CPU_CORE_UTILIZATION, 
+           ProcessMonitorKey.MEMORY_RESIDENT_KB,
+           ProcessMonitorKey.MEMORY_VIRTUAL_KB,
+        ], writeHeaderLine=False)